change app title and add readme
